✨ feat(posts): integrate S3 storage for images, enhance post creation with image upload validation, and update Swagger documentation for improved API clarity

+++ app/controllers/posts_controller.ts
@@ -2,6 +2,10 @@ import type { HttpContext } from '@adonisjs/core/http'
 import Post from '#models/post'
 import PostComment from '#models/post_comment'
 import { createPostValidator } from '#validators/create_post'
+import { createPostCommentValidator } from '#validators/create_post_comment'
+import { v4 as uuid } from 'uuid'
+import drive from '@adonisjs/drive/services/main'
+import { readFile } from 'node:fs/promises'
 
 export default class PostsController {
   /**
@@ -20,47 +24,51 @@ export default class PostsController {
    * @store
    * @description Create a new post
    * @requestBody <createPostValidator>
-   * @responseBody 200 - <Post>
+   * @responseBody 201 - <Post>
    */
   public async store({ auth, request, response }: HttpContext) {
-    const payload = await createPostValidator.validate(request.only(['content']))
+    const post = new Post()
     const user = auth.getUserOrFail()
 
-    const post = await Post.create({ content: payload.content, userId: user.id, likes: 0 })
+    const { content, image } = await request.validateUsing(createPostValidator)
+
+    post.fill({
+      content: content,
+      userId: user.id,
+    })
+
+    if (image) {
+      const fileName = `${uuid()}.${image.extname}`
+      const fileBuffer = await readFile(image.tmpPath!)
+      await drive.use('s3').put(fileName, fileBuffer, {
+        contentType: image.type,
+        visibility: 'public',
+      })
+
+      post.image = await drive.use('s3').getUrl(fileName)
+    }
+
+    await post.save()
     return response.json(post)
   }
 
   /**
    * @show
    * @description Get a post by id
-   * @requestBody <Post>
-   * @responseBody 200 - <Post>
+   * @responseBody 200 - <Post>.with(comments)
    */
   public async show({ request, response }: HttpContext) {
     const { id } = request.params()
     const post = await Post.find(id)
-    return response.json(post)
-  }
 
-  /**
-   * @update
-   * @description Update a post by id
-   * @requestBody <Post>
-   * @responseBody 200 - <Post>
-   */
-  public async update({ request, response }: HttpContext) {
-    const { id } = request.params()
-    const { content } = request.only(['content'])
+    const comments = await PostComment.query().where('post_id', id)
 
-    const post = await Post.find(id)
-
-    if (!post) {
-      return response.status(404).json({ message: 'Post not found' })
+    const postWithComments = {
+      ...post,
+      comments,
     }
 
-    post.content = content
-    await post.save()
-    return response.json(post)
+    return response.json(postWithComments)
   }
 
   /**
@@ -69,12 +77,18 @@ export default class PostsController {
    * @requestBody <Post>
    * @responseBody 200 - <Post>
    */
-  public async destroy({ request, response }: HttpContext) {
+  public async destroy({ auth, request, response }: HttpContext) {
+    const user = auth.getUserOrFail()
+
     const { id } = request.params()
     const post = await Post.find(id)
 
     if (!post) {
       return response.status(404).json({ message: 'Post not found' })
+    }
+
+    if (post.userId !== user.id) {
+      return response.status(403).json({ message: 'You are not authorized to delete this post' })
     }
 
     await post.delete()
@@ -83,6 +97,7 @@ export default class PostsController {
 
   /**
    * @like
+   * @summary Like a post by id
    * @description Like a post by id
    * @requestBody <Post>
    * @responseBody 200 - <Post>
@@ -102,13 +117,15 @@ export default class PostsController {
 
   /**
    * @comment
+   * @summary Comment on a post by id
    * @description Comment on a post by id
-   * @requestBody <Post>
-   * @responseBody 200 - <Post>
+   * @requestBody <createPostCommentValidator>
+   * @responseBody 201 - <Post>
    */
   public async comment({ auth, request, response, logger }: HttpContext) {
     const { id } = request.params()
-    const { content } = request.only(['content'])
+    const payload = await createPostCommentValidator.validate(request.only(['content']))
+    const content = payload.content
 
     logger.info({ id, content }, 'Commenting on post')
